Update sample_gamma_eta.cpp and initialize analyze_sample_gamma_eta.R
I've now got a main function in sample_gamma_eta.cpp in that it writes outputs of the desired structure. On initializing analyze_sample_gamma_eta.R to visualize the trace plots, the Gibbs mixing seems to be working, but there's seemingly an issue in the metropolis-hastings step. The big win is that the chain now requires 2.45 minutes to run 6,000 iterations.