Update RcppArmadillo.h & RcppGSL.h include order
There was an issue when RcppGSL.h was included before RcppArmadillo.h. This was because RcppGSL.h loads Rcpp.h, and RcppArmadillo.h throws an error if Rcpp.h has been loaded previously. This is a workaround that might be alleviated eventually by converting to only RcppArmadillo.h over time.