Update BIP_with_mixed_model_embed.cpp and gibbs_sampler_nested.cpp
These updates include a successful embedding of the mixed outcome model into the BIP framework. A challenge is apparent in that the sigma2 is overestimated for the outcome view, and sigma2_ksi is underestimated. This is likely attributable to the fact that each is reliant on a residualization that induces autocorrelation in the sampling of each parameter. TBD how to resolve this issue, though sampling the outcome model's variance parameter independent of the other views' sigma2s seems a promising direction.